+++ packages/core/src/watch/query-hooks/types.ts
@@ -1,8 +1,13 @@
+import { transformPlaylistFragment, transformVideoFragment } from './transformers';
+
 const MEDIA_TYPES = {
   VIDEO: 'video',
   PLAYLIST: 'playlist',
 } as const;
 
 type MediaType = (typeof MEDIA_TYPES)[keyof typeof MEDIA_TYPES];
+type TransformedVideo = ReturnType<typeof transformVideoFragment>;
+type TransformedPlaylist = ReturnType<typeof transformPlaylistFragment>;
+type TransformedMediaItem = TransformedVideo | TransformedPlaylist;
 
-export { type MediaType, MEDIA_TYPES };
+export { MEDIA_TYPES, type MediaType, type TransformedVideo, type TransformedPlaylist, type TransformedMediaItem };
