+++ packages/core/src/watch/query-hooks/videos/index.ts
@@ -1,10 +1,7 @@
-import { FragmentType, getFragmentData, graphql } from '../../../graphql';
+import { graphql } from '../../../graphql';
 import { AllVideosQuery } from '../../../graphql/graphql';
-import { AppError } from '../../../universal/error-boundary/app-error';
 import { useRequest } from '../../../universal/hooks/use-request';
-import { PlaylistFragment, PlaylistVideoFragment, UserFragment, VideoFragment } from '../fragments';
-import { transformVideoFragment } from '../transformers';
-import { MEDIA_TYPES } from '../types';
+import { transformPlaylistFragment, transformVideoFragment } from '../transformers';
 
 const videosQuery = graphql(/* GraphQL */ `
   query AllVideos @cached {
@@ -23,58 +20,6 @@ const videosQuery = graphql(/* GraphQL */ `
 interface LoadVideosProps {
   getAccessToken: () => Promise<string>;
 }
-
-// interface BaseVideoInfo {
-//   id: string;
-//   title: string;
-//   description: string;
-//   thumbnailUrl: string;
-//   slug: string;
-//   createdAt: string;
-//   user: User | null;
-//   type: MediaType;
-// }
-
-// interface TransformedVideo extends BaseVideoInfo {
-//   source: string;
-//   duration: number;
-//   lastWatchedAt: string | null;
-//   progressSeconds: number;
-// }
-
-// interface TransformedPlaylist extends BaseVideoInfo {
-//   firstVideoId: string;
-// }
-
-type TransformedVideo = ReturnType<typeof transformVideoFragment>;
-type TransformedPlaylist = ReturnType<typeof transformPlaylistFragment>;
-type TransformedMediaItem = TransformedVideo | TransformedPlaylist;
-
-const transformPlaylistFragment = (playlistFragmentData: FragmentType<typeof PlaylistFragment>) => {
-  const playlist = getFragmentData(PlaylistFragment, playlistFragmentData);
-  const user = {
-    username: getFragmentData(UserFragment, playlist.user).username || '',
-  };
-
-  if (!playlist.playlist_videos.length) {
-    throw new AppError('Playlist has no videos', 'Playlist has no videos', false);
-  }
-
-  const firstPlaylistVideo = getFragmentData(PlaylistVideoFragment, playlist.playlist_videos[0]);
-  const firstVideo = getFragmentData(VideoFragment, firstPlaylistVideo.video);
-
-  return {
-    id: playlist.id,
-    type: MEDIA_TYPES.PLAYLIST,
-    title: playlist.title,
-    thumbnailUrl: playlist.thumbnailUrl,
-    slug: playlist.slug,
-    createdAt: playlist.createdAt,
-    description: playlist.description || '',
-    user,
-    firstVideoId: firstVideo.id,
-  };
-};
 
 const transform = (data: AllVideosQuery) => {
   const { videos, playlist } = data;
@@ -104,4 +49,4 @@ const useLoadVideos = (props: LoadVideosProps) => {
   };
 };
 
-export { type TransformedVideo, type TransformedPlaylist, type TransformedMediaItem, useLoadVideos };
+export { useLoadVideos };
