+++ packages/ui/src/watch/video-detail-page/containers/types.ts
@@ -1,0 +1,8 @@
+import { useLoadPlaylistDetail } from 'core/watch/query-hooks/playlist-detail';
+import { RequiredLinkComponent } from '../../videos/types';
+
+interface VideoDetailContainerProps extends Omit<RequiredLinkComponent, 'linkProps'> {
+  queryRs: ReturnType<typeof useLoadPlaylistDetail>;
+}
+
+export { type VideoDetailContainerProps };
